Updating PInvokeGenerator to use the true kind when visiting cursors
