+++ sources/ClangSharp.PInvokeGenerator/PInvokeGenerator.VisitAttr.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft and Contributors. All rights reserved. Licensed under the University of Illinois/NCSA Open Source License. See LICENSE.txt in the project root for license information.
+
+namespace ClangSharp
+{
+    public partial class PInvokeGenerator
+    {
+        private void VisitAttr(Attr attr)
+        {
+            // We don't consider most attributes particularly important and so we do nothing
+        }
+    }
+}
