+++ sources/ClangSharp/Cursors/Decls/LinkageSpecDecl.cs
@@ -23,6 +23,8 @@ namespace ClangSharp
 
         public IReadOnlyList<Decl> Decls => _decls.Value;
 
+        public CXLanguageKind Langage => Handle.Language;
+
         public IDeclContext LexicalParent => LexicalDeclContext;
 
         public IDeclContext Parent => DeclContext;
